requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
python-socketio==5.11.0
python-engineio==4.8.0
python-dotenv==1.0.1
Werkzeug==3.0.1
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
gunicorn==21.2.0
eventlet==0.33.3
pandas==2.1.4
numpy==1.26.3
pytz==2023.3.post1
Jinja2==3.1.3
MarkupSafe==2.1.5
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
XlsxWriter==3.1.9
python-jose==3.3.0
cryptography==41.0.7
requests==2.31.0
redis==5.0.1
beautifulsoup4==4.12.2 